fix: sim CFN Fn::Sub filter implicit references with full variables
